Add an interval between connection attempts
For connection attempts to the same remote device, add 5 seconds
interval until the next connect attempt. This setting effects mainly
address not found in the kernel advertising cache: LE scanning will
be executed in background during a short period of time until it finds
the address or EHOSTDOWN is returned to the connection attempt.